\name{Sweparl-package}
\alias{Sweparl-package}
\alias{Sweparl}
\docType{package}
\title{
\packageTitle{Sweparl}
}
\description{
\packageDescription{Sweparl}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Sweparl}
\packageIndices{Sweparl}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Sweparl}

Maintainer: \packageMaintainer{Sweparl}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
