% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swesick.R
\name{swesick}
\alias{swesick}
\title{Download data for quarterly ongoing illnesses in Sweden}
\usage{
swesick(year = NULL)
}
\arguments{
\item{year}{Optional vector to download data for specific years.}
}
\value{
A \code{data.frame} with the downloaded data.
}
\description{
The function downloads data for quarterly ongoing illnesses in Sweden from 2005 and onward. The data is provided by the API for the Swedish Social Insurance Agency (Forsakringskassan).
}
\examples{
sickdata <- swesick()
sickdata_0506 <- swesick(year = c(2005,2006))
}
\references{
Swedish Social Insurance Agency. \url{https://catalog.forsakringskassan.se/catalog/9/datasets/21}
}
